int -> size_t

+++ grechishnikov.roman/P3/countUnique.cpp
@@ -10,7 +10,7 @@ size_t grechishnikov::countUnique(const char* str)
     {
       char del = std::tolower(str[i]);
       int check = 1;
-      for (int g = 0; g < i; g++)
+      for (size_t g = 0; g < i; g++)
       {
         if (del == std::tolower(str[g]))
         {
